Split up the backend, frontend, and the docprovider
