+++ packages/docprovider-extension/README.md
@@ -1,0 +1,3 @@
+# @jupyter/docprovider-extension
+
+A JupyterLab package which provides a set of plugins for collaborative shared models.
