+++ packages/docprovider-extension/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyter/collaboration-extension",
+  "name": "@jupyter/docprovider-extension",
   "version": "2.0.11",
-  "description": "JupyterLab - Real-Time Collaboration Extension",
+  "description": "JupyterLab - Collaborative Shared Models",
   "keywords": [
     "jupyter",
     "jupyterlab",
@@ -45,7 +45,7 @@
     "build:labextension:dev": "jupyter labextension build --development True .",
     "clean": "jlpm run clean:lib",
     "clean:lib": "rimraf lib tsconfig.tsbuildinfo node_modules",
-    "clean:labextension": "rimraf ../../projects/jupyter-collaboration-ui/jupyter_collaboration_ui/labextension",
+    "clean:labextension": "rimraf ../../projects/jupyter-docprovider/jupyter_docprovider/labextension",
     "clean:all": "jlpm run clean:lib && jlpm run clean:labextension",
     "install:extension": "jlpm run build",
     "watch": "run-p watch:src watch:labextension",
@@ -53,18 +53,18 @@
     "watch:labextension": "jupyter labextension watch ."
   },
   "dependencies": {
-    "@jupyter/collaboration": "^2.0.11",
     "@jupyter/docprovider": "^2.0.11",
     "@jupyter/ydoc": "^1.1.0-a0",
     "@jupyterlab/application": "^4.2.0-beta.0",
     "@jupyterlab/apputils": "^4.2.0-beta.0",
-    "@jupyterlab/codemirror": "^4.2.0-beta.0",
-    "@jupyterlab/coreutils": "^6.0.5",
-    "@jupyterlab/services": "^7.0.5",
-    "@jupyterlab/statedb": "^4.2.0-beta.0",
+    "@jupyterlab/docregistry": "^4.2.0-beta.0",
+    "@jupyterlab/filebrowser": "^4.2.0-beta.0",
+    "@jupyterlab/fileeditor": "^4.2.0-beta.0",
+    "@jupyterlab/logconsole": "^4.2.0-beta.0",
+    "@jupyterlab/notebook": "^4.2.0-beta.0",
+    "@jupyterlab/settingregistry": "^4.2.0-beta.0",
     "@jupyterlab/translation": "^4.2.0-beta.0",
-    "@jupyterlab/ui-components": "^4.2.0-beta.0",
-    "@lumino/widgets": "^2.1.0",
+    "@lumino/commands": "^2.1.0",
     "y-protocols": "^1.0.5",
     "y-websocket": "^1.3.15",
     "yjs": "^13.5.40"
@@ -82,13 +82,16 @@
   "typedoc": {
     "entryPoint": "./src/index.ts",
     "readmeFile": "./README.md",
-    "displayName": "@jupyter/collaboration-extension",
+    "displayName": "@jupyter/docprovider-extension",
     "tsconfig": "./tsconfig.json"
   },
   "jupyterlab": {
     "extension": true,
-    "schemaDir": "./schema",
-    "outputDir": "../../projects/jupyter-collaboration-ui/jupyter_collaboration_ui/labextension",
+    "outputDir": "../../projects/jupyter-docprovider/jupyter_docprovider/labextension",
+    "disabledExtensions": [
+      "@jupyterlab/filebrowser-extension:defaultFileBrowser",
+      "@jupyterlab/notebook-extension:cell-executor"
+    ],
     "sharedPackages": {
       "@codemirror/state": {
         "bundled": false,
@@ -96,10 +99,6 @@
       },
       "@codemirror/view": {
         "bundled": false,
-        "singleton": true
-      },
-      "@jupyter/collaboration": {
-        "bundled": true,
         "singleton": true
       },
       "@jupyter/docprovider": {
