+++ packages/docprovider-extension/style/index.css
@@ -1,0 +1,6 @@
+/* -----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|---------------------------------------------------------------------------- */
+
+@import url('~@jupyter/collaboration/style/index.css');
