+++ packages/docprovider-extension/style/index.js
@@ -1,0 +1,6 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+import '@jupyter/collaboration/style/index.js';
