+++ projects/jupyter-collaboration-ui/README.md
@@ -1,0 +1,3 @@
+# jupyter-collaboration-ui
+
+JupyterLab/Jupyter Notebook 7+ extension providing user interface integration for real time collaboration.
