+++ projects/jupyter-collaboration-ui/install.json
@@ -1,0 +1,5 @@
+{
+  "packageManager": "python",
+  "packageName": "jupyter-collaboration-ui",
+  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package jupyter-collaboration-ui"
+}
