+++ projects/jupyter-collaboration/README.md
@@ -1,0 +1,6 @@
+# jupyter-collaboration
+
+A meta-package for:
+- jupyter-collaboration-ui
+- jupyter-docprovider
+- jupyter-server-ydoc
