+++ projects/jupyter-collaboration/jupyter_collaboration/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
+
+from ._version import __version__  # noqa
