+++ projects/jupyter-collaboration/jupyter_collaboration/_version.py
@@ -1,0 +1,1 @@
+__version__ = "2.0.11"
