+++ projects/jupyter-docprovider/README.md
@@ -1,0 +1,7 @@
+# jupyter-docprovider
+
+JupyterLab/Jupyter Notebook 7+ extension integrating collaborative shared models.
+
+The collaborative shared models are used for both:
+- real time collaboration, and
+- server-side execution of notebooks
