+++ projects/jupyter-docprovider/install.json
@@ -1,0 +1,5 @@
+{
+  "packageManager": "python",
+  "packageName": "jupyter-docprovider",
+  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package jupyter-docprovider"
+}
