+++ projects/jupyter-server-ydoc/jupyter-config/jupyter_server_ydoc.json
@@ -1,7 +1,7 @@
 {
   "ServerApp": {
     "jpserver_extensions": {
-      "jupyter_collaboration": true
+      "jupyter_server_ydoc": true
     }
   }
 }
