+++ projects/jupyter-server-ydoc/jupyter_server_ydoc/__init__.py
@@ -7,9 +7,5 @@ from ._version import __version__  # noqa
 from .app import YDocExtension
 
 
-def _jupyter_labextension_paths():
-    return [{"src": "labextension", "dest": "@jupyter/collaboration-extension"}]
-
-
 def _jupyter_server_extension_points() -> List[Dict[str, Any]]:
     return [{"module": "jupyter_collaboration", "app": YDocExtension}]
