+++ projects/jupyter-server-ydoc/pyproject.toml
@@ -1,0 +1,49 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
+
+[build-system]
+build-backend = "hatchling.build"
+requires = ["hatchling>=1.4.0"]
+
+[project]
+name = "jupyter-server-ydoc"
+readme = "README.md"
+license = { file = "../../LICENSE" }
+requires-python = ">=3.8"
+description = "jupyter-server extension integrating collaborative shared models."
+classifiers = [
+    "Intended Audience :: Developers",
+    "Intended Audience :: System Administrators",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: BSD License",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Framework :: Jupyter"
+]
+authors = [
+    { name = "Jupyter Development Team", email = "jupyter@googlegroups.com" },
+]
+dependencies = [
+    "jupyter_server>=2.4.0,<3.0.0",
+    "jupyter_ydoc>=2.0.0,<3.0.0",
+    "pycrdt-websocket>=0.12.5,<0.13.0",
+    "jupyter_events>=0.10.0",
+    "jupyter_server_fileid>=0.7.0,<1",
+    "jsonschema>=4.18.0"
+]
+dynamic = ["version"]
+
+[tool.hatch.version]
+path = "jupyter_server_ydoc/_version.py"
+
+[tool.hatch.build.targets.sdist]
+exclude = ["/.github", "/binder", "node_modules"]
+
+[tool.hatch.build.targets.wheel.shared-data]
+"jupyter-config/jupyter_server_ydoc.json" = "etc/jupyter/jupyter_server_config.d/jupyter_collaboration.json"
+
+[tool.check-wheel-contents]
+ignore = ["W002"]
