+++ tests/test_app.py
@@ -4,12 +4,11 @@
 from __future__ import annotations
 
 import pytest
-
-from jupyter_collaboration.stores import SQLiteYStore, TempFileYStore
+from jupyter_server_ydoc.stores import SQLiteYStore, TempFileYStore
 
 
 def test_default_settings(jp_serverapp):
-    settings = jp_serverapp.web_app.settings["jupyter_collaboration_config"]
+    settings = jp_serverapp.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["disable_rtc"] is False
     assert settings["file_poll_interval"] == 1
@@ -22,7 +21,7 @@ def test_settings_should_disable_rtc(jp_configurable_serverapp):
     argv = ["--YDocExtension.disable_rtc=True"]
 
     app = jp_configurable_serverapp(argv=argv)
-    settings = app.web_app.settings["jupyter_collaboration_config"]
+    settings = app.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["disable_rtc"] is True
 
@@ -31,7 +30,7 @@ def test_settings_should_change_file_poll(jp_configurable_serverapp):
     argv = ["--YDocExtension.file_poll_interval=2"]
 
     app = jp_configurable_serverapp(argv=argv)
-    settings = app.web_app.settings["jupyter_collaboration_config"]
+    settings = app.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["file_poll_interval"] == 2
 
@@ -40,7 +39,7 @@ def test_settings_should_change_document_cleanup(jp_configurable_serverapp):
     argv = ["--YDocExtension.document_cleanup_delay=10"]
 
     app = jp_configurable_serverapp(argv=argv)
-    settings = app.web_app.settings["jupyter_collaboration_config"]
+    settings = app.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["document_cleanup_delay"] == 10
 
@@ -49,16 +48,16 @@ def test_settings_should_change_save_delay(jp_configurable_serverapp):
     argv = ["--YDocExtension.document_save_delay=10"]
 
     app = jp_configurable_serverapp(argv=argv)
-    settings = app.web_app.settings["jupyter_collaboration_config"]
+    settings = app.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["document_save_delay"] == 10
 
 
 def test_settings_should_change_ystore_class(jp_configurable_serverapp):
-    argv = ["--YDocExtension.ystore_class=jupyter_collaboration.stores.TempFileYStore"]
+    argv = ["--YDocExtension.ystore_class=jupyter_server_ydoc.stores.TempFileYStore"]
 
     app = jp_configurable_serverapp(argv=argv)
-    settings = app.web_app.settings["jupyter_collaboration_config"]
+    settings = app.web_app.settings["jupyter_server_ydoc_config"]
 
     assert settings["ystore_class"] == TempFileYStore
 
@@ -66,7 +65,7 @@ def test_settings_should_change_ystore_class(jp_configurable_serverapp):
 @pytest.mark.parametrize("copy", [True, False])
 async def test_get_document_file(rtc_create_file, jp_serverapp, copy):
     path, content = await rtc_create_file("test.txt", "test", store=True)
-    collaboration = jp_serverapp.web_app.settings["jupyter_collaboration"]
+    collaboration = jp_serverapp.web_app.settings["jupyter_server_ydoc"]
     document = await collaboration.get_document(
         path=path, content_type="file", file_format="text", copy=copy
     )
@@ -78,7 +77,7 @@ async def test_get_document_file_copy_is_independent(
     rtc_create_file, jp_serverapp, rtc_fetch_session
 ):
     path, content = await rtc_create_file("test.txt", "test", store=True)
-    collaboration = jp_serverapp.web_app.settings["jupyter_collaboration"]
+    collaboration = jp_serverapp.web_app.settings["jupyter_server_ydoc"]
     document = await collaboration.get_document(
         path=path, content_type="file", file_format="text", copy=True
     )
